Use malloc()-based allocator
